refactor ui files
