+++ UI/src/lib/components/products/edit/create.edit.price.svelte
@@ -1,0 +1,25 @@
+<script lang="ts">
+	export let product_price: number;
+</script>
+
+<input
+	type="number"
+	class="form-control
+										  block
+										  w-full
+										  px-3
+										  py-1.5
+										  text-base
+										  font-normal
+										  text-gray-700
+										  bg-white bg-clip-padding
+										  border border-solid border-gray-300
+										  rounded
+										  transition
+										  ease-in-out
+										  m-0
+										  focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none mb-2 "
+	id="edit_product_price"
+	placeholder="Price"
+	bind:value={product_price}
+/>
