+++ UI/src/lib/components/products/product.card.svelte
@@ -2,8 +2,8 @@
 	import type { Product } from '$lib/models/product';
 	import ProductTypeIconComponent from '$lib/components/products/product.type.icon.svelte';
 	import ProductArchive from '$lib/components/products/product.archive.svelte';
-	import ProductRanking from '$lib/components/products/product.ranking.svelte';
-	import ProductEdit from '$lib/components/products/product.edit.svelte';
+	import ProductRanking from '$lib/components/products/ranking/view.product.ranking.svelte';
+	import ProductEdit from '$lib/components/products/edit/product.edit.svelte';
 
     export let product: Product;
 	
@@ -67,7 +67,6 @@
 				>
 					<path
 						stroke-linecap="round"
-						stroke-linejoin="round"
 						d="M20.25 7.5l-.625 10.632a2.25 2.25 0 01-2.247 2.118H6.622a2.25 2.25 0 01-2.247-2.118L3.75 7.5m6 4.125l2.25 2.25m0 0l2.25 2.25M12 13.875l2.25-2.25M12 13.875l-2.25 2.25M3.375 7.5h17.25c.621 0 1.125-.504 1.125-1.125v-1.5c0-.621-.504-1.125-1.125-1.125H3.375c-.621 0-1.125.504-1.125 1.125v1.5c0 .621.504 1.125 1.125 1.125z"
 					/>
 				</svg>
